requirements.txt
crewai>=0.28.0
langchain>=0.2.0
langchain-ollama>=0.1.0
langchain-core>=0.2.0
web3>=6.0.0
python-dotenv>=1.0.0
rich>=13.0.0
